Update and rename file4_Types_of_critical_points.py to file4_Relative_Maxima_and_Relative_Minima.py
